Initial layout and gatsby file setup.
Organised project into a monorepo with the following structure:
```
packages
├ gatsby-theme-portfolio
└ www
```

I set up Lerna to allow me to use the theme from within the same project. This means the content can be completely independent from the theme. The plan is to publish the theme.

The theme is based on `gatsby-theme-ui-blog` and `gatsby-plugin-theme-ui` to allow easy customisation of the theme and colorscheme.

This is still a work in progress and so there are plenty of things left unfinished.